Moved CharacteristicEntry component to characteristicEntry.jsx for modularity
